p4i integration Integrate 204021 from 4.1 to main: Fixes:    win32 compile.. RevBy:    trustme.. Details:
[git-p4: depot-paths = "//depot/qt/main/": change = 204022]

+++ doc/src/snippets/qtablewidget-dnd/mainwindow.cpp
@@ -51,7 +51,7 @@ void MainWindow::setupTableItems()
     for (int row = 0; row < tableWidget->rowCount()-1; ++row) {
         for (int column = 0; column < tableWidget->columnCount(); ++column) {
     QTableWidgetItem *newItem = new QTableWidgetItem(tr("%1").arg(
-        pow(row, column+1)));
+        pow((float)row, (float)column+1)));
     tableWidget->setItem(row, column, newItem);
         }
     }
